requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backoff==2.2.1
beautifulsoup4==4.12.0
bs4==0.0.1
CacheControl==0.12.11
caio==0.9.12
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==4.4.2
frozenlist==1.3.3
idna==3.4
imageio-ffmpeg==0.4.8
imageio==2.26.1
magic-filter==1.0.9
moviepy==1.0.3
msgpack==1.0.5
multidict==6.0.4
numpy==1.24.2
Pillow==9.4.0
pip==22.3.1
proglog==0.1.10
pyaes==1.6.1
pyasn1==0.4.8
PySocks==1.7.1
pytube==12.1.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
setuptools==65.5.1
soupsieve==2.4
Telethon==1.27.0
tqdm==4.65.0
urllib3==1.26.15
wheel==0.38.4
yarl==1.8.2
youtube-transcript-api==0.5.0
